requirements.txt
httpx==0.25.0
python-dotenv==1.0.0
pytest==7.4.2
